# src/pwa/public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "WebAZ Protocol API",
    "version": "0.4.14",
    "description": "Auto-generated endpoint inventory (640 endpoints, 17 with full schema). See docs/ for design context."
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "api_key"
      },
      "webauthnToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-WebAuthn-Token"
      }
    }
  },
  "paths": {
...
    "/api/rewards/apply": {
      "post": {
        "summary": "POST /api/rewards/apply — activate (or reconfirm) opt-in + drain escrow",
        "tags": [],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/rewards/deactivate": {
      "post": {
        "summary": "POST /api/rewards/deactivate — flip off; subsequent commissions → charity",
        "tags": [],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/rewards/status": {
      "get": {
        "summary": "GET /api/rewards/status — current state + escrow tally",
        "tags": [],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  }
}